Move Cta component to ui folder
