+++ components/headers/ExperimentHeader.vue
@@ -39,7 +39,7 @@
 <script lang="ts">
 import Vue from 'vue'
 import { Component, Prop } from 'vue-property-decorator'
-import Cta from '~/components/ctas/Cta.vue'
+import Cta from '~/components/ui/Cta.vue'
 import Media from '~/components/Media.vue'
 
 @Component({
